+++ project-base/src/Model/Article/Article.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Model\Article;
 
-use DateTime;
 use Doctrine\ORM\Mapping as ORM;
 use Shopsys\FrameworkBundle\Model\Article\Article as BaseArticle;
 use Shopsys\FrameworkBundle\Model\Article\ArticleData as BaseArticleData;
@@ -15,12 +14,6 @@ use Shopsys\FrameworkBundle\Model\Article\ArticleData as BaseArticleData;
  */
 class Article extends BaseArticle
 {
-    /**
-     * @var \DateTime
-     * @ORM\Column(type="datetime")
-     */
-    private $createdAt;
-
     /**
      * @param \App\Model\Article\ArticleData $articleData
      */
@@ -43,15 +36,5 @@ class Article extends BaseArticle
     protected function setData(BaseArticleData $articleData): void
     {
         parent::setData($articleData);
-
-        $this->createdAt = $articleData->createdAt ?? new DateTime();
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getCreatedAt(): DateTime
-    {
-        return $this->createdAt;
     }
 }
