+++ project-base/src/Model/Article/ArticleDataFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Model\Article;
 
+use Shopsys\FrameworkBundle\Model\Article\Article;
 use Shopsys\FrameworkBundle\Model\Article\ArticleData as BaseArticleData;
 use Shopsys\FrameworkBundle\Model\Article\ArticleDataFactory as BaseArticleDataFactory;
 
@@ -19,5 +20,14 @@ class ArticleDataFactory extends BaseArticleDataFactory
     protected function createInstance(): BaseArticleData
     {
         return new ArticleData();
+    }
+
+    /**
+     * @param \App\Model\Article\ArticleData $articleData
+     * @param \App\Model\Article\Article $article
+     */
+    protected function fillFromArticle(BaseArticleData $articleData, Article $article): void
+    {
+        parent::fillFromArticle($articleData, $article);
     }
 }
