+++ symfony.lock
@@ -428,9 +428,6 @@
     "phpdocumentor/type-resolver": {
         "version": "0.4.0"
     },
-    "phpspec/prophecy": {
-        "version": "1.9.0"
-    },
     "phpstan/phpdoc-parser": {
         "version": "0.3.5"
     },
